+++ src/api/RealtimeAPI.ts
@@ -1,6 +1,7 @@
 import WebSocketAPI, {
   WebSocketAPIMessageCallback,
   WebSocketAPIMessageEventData,
+  WebSocketAPIParameters,
 } from '../common/api/WebSocketAPI';
 import debounceWebsocketMessages from '../common/utils/debounceWebsocketMessages';
 import getModeSuffix from '../common/utils/getRealtimeModeSuffix';
@@ -16,6 +17,7 @@ import type {
   RealtimeTrajectoryResponse,
   RealtimeStationId,
   RealtimeVersion,
+  RealtimeTrajectory,
 } from '../types';
 import { StopSequence } from './typedefs';
 
@@ -111,10 +113,10 @@ class RealtimeAPI {
   constructor(options: RealtimeAPIOptions = {}) {
     this.defineProperties(options);
 
-    /** @ignore */
+    /** @private */
     this.prefix = options.prefix || '';
 
-    /** @ignore */
+    /** @private */
     this.onOpen = this.onOpen.bind(this);
   }
 
@@ -271,7 +273,7 @@ class RealtimeAPI {
      */
     if (this.pingIntervalMs) {
       window.clearInterval(this.pingInterval);
-      /** @ignore */
+      /** @private */
       this.pingInterval = window.setInterval(() => {
         this.wsApi.send('PING');
       }, this.pingIntervalMs);
@@ -287,12 +289,32 @@ class RealtimeAPI {
     window.clearTimeout(this.reconnectTimeout);
 
     if (this.reconnectTimeoutMs) {
-      /** @ignore */
+      /** @private */
       this.reconnectTimeout = window.setTimeout(
         () => this.open(),
         this.reconnectTimeoutMs,
       );
     }
+  }
+
+  /**
+   * Send GET to a channel.
+   *
+   * @param {string | WebSocketAPIParameters} channelOrParams Name of the websocket channel to send GET or an object representing parameters to send
+   * @return {Promise<WebSocketAPIMessageEventData<?>>} A websocket response.
+   */
+  get(
+    channelOrParams: string | WebSocketAPIParameters,
+  ): Promise<WebSocketAPIMessageEventData<any>> {
+    let params = channelOrParams as WebSocketAPIParameters;
+
+    if (typeof channelOrParams === 'string') {
+      params = { channel: channelOrParams };
+    }
+
+    return new Promise((resolve, reject) => {
+      this.wsApi.get(params, resolve, reject);
+    });
   }
 
   /**
@@ -419,9 +441,7 @@ class RealtimeAPI {
       args: uic,
     };
 
-    return new Promise((resolve, reject) => {
-      this.wsApi.get(params, resolve, reject);
-    });
+    return this.get(params);
   }
 
   /**
@@ -432,10 +452,7 @@ class RealtimeAPI {
    */
   getStations(mode: RealtimeMode, timeout = 100): Promise<RealtimeStation[]> {
     return new Promise((resolve) => {
-      this.wsApi.get(
-        {
-          channel: `station${getModeSuffix(mode, RealtimeModes)}`,
-        },
+      this.get(`station${getModeSuffix(mode, RealtimeModes)}`).then(
         debounceWebsocketMessages(resolve, undefined, timeout),
       );
     });
@@ -497,6 +514,22 @@ class RealtimeAPI {
     onMessage: WebSocketAPIMessageCallback<RealtimeExtraGeom>,
   ) {
     this.unsubscribe('extra_geoms', '', onMessage);
+  }
+
+  /**
+   * Return a partial trajectory with a given id and a mode.
+   *
+   * @param {number} trainId The identifier of a trajectory.
+   * @param {RealtimeMode} mode Realtime mode.
+   * @return {Promise<{data: { content: RealtimeTrajectory }}>} A trajectory.
+   */
+  getTrajectory(
+    id: RealtimeTrainId,
+    mode: RealtimeMode,
+  ): Promise<WebSocketAPIMessageEventData<RealtimeTrajectory>> {
+    return this.get(
+      `partial_trajectory${getModeSuffix(mode, RealtimeModes)}_${id}`,
+    );
   }
 
   /**
@@ -575,7 +608,7 @@ class RealtimeAPI {
    * @param {string} id A vehicle id.
    * @param {RealtimeMode} mode Realtime mode.
    * @param {string} generalizationLevel The generalization level to request. Can be one of 5 (more generalized), 10, 30, 100, undefined (less generalized).
-   * @return {Promise<{ data: { content: FullTrajectory } }>} Return a full trajectory.
+   * @return {Promise<{ data: { content: RealtimeFullTrajectory } }>} Return a full trajectory.
    */
   getFullTrajectory(
     id: RealtimeTrainId,
@@ -596,13 +629,7 @@ class RealtimeAPI {
       channel.push(`gen${generalizationLevel}`);
     }
 
-    const params = {
-      channel: channel.join('_'),
-    };
-
-    return new Promise((resolve, reject) => {
-      this.wsApi.get(params, resolve, reject);
-    });
+    return this.get(channel.join('_'));
   }
 
   /**
@@ -651,15 +678,7 @@ class RealtimeAPI {
   getStopSequence(
     id: RealtimeTrainId,
   ): Promise<WebSocketAPIMessageEventData<StopSequence[]>> {
-    return new Promise((resolve, reject) => {
-      this.wsApi.get(
-        {
-          channel: `stopsequence_${id}`,
-        },
-        resolve,
-        reject,
-      );
-    });
+    return this.get(`stopsequence_${id}`);
   }
 
   /**